fix(postgres): pressing Enter on a row panics when opening details modal
The user story was: opening row details on a PostgreSQL database
by pressing Enter on a table row caused a panic instead of showing
the details modal.

Root cause: the get_foreign_keys() query selected con.oid as a
PostgreSQL oid type. The postgres crate deserializes oid as u32,
but the code used row.get::<_, i32>(0), which panics at runtime
when the type mismatch is detected.

Fix: read con.oid as u32 and cast safely to i32 for the
ForeignKeyInfo.id field.

+++ src/driver/postgres.rs
@@ -251,7 +251,7 @@ impl DbDriver for PostgresDriver {
         let mut fks = Vec::new();
         for row in rows {
             fks.push(ForeignKeyInfo {
-                id: row.get::<_, i32>(0),
+                id: row.get::<_, u32>(0) as i32,
                 seq: (row.get::<_, i64>(4) - 1) as i32,
                 from: row.get::<_, String>(1),
                 table: row.get::<_, String>(2),
